backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
python-multipart
alembic
pydantic-settings
email-validator
bcrypt==4.3.0